requirements.txt
streamlit
requests
pillow
matplotlib
ipython
openai
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
faiss-cpu